+++ dts/interfaces.d.ts
@@ -24,9 +24,30 @@ export declare type ServiceFactory<T = any> = (context: Context) => T;
  * An object capable of resolving identifiers to objects.
  */
 export interface ServiceLocator {
-    create<T>(ctor: Newable<T>): T;
-    get<T>(identifier: Identifier<T>): T;
+    /**
+     * Creates a new instance of the class, resolving its injections.
+     * @param ctor The class constructor to instantiate.
+     * @param parameters Parameters to pass to parameter injections.
+     */
+    create<T>(ctor: Newable<T>, parameters?: ParameterRecord): T;
+    /**
+     * Resolves an identifier from the container bindings.
+     * @param identifier The identifier to resolve.
+     * @param parameters Parameters to pass to parameter injections for any new objects created.
+     *
+     * Note: Parameters will not apply if the object or objects being resolved have already
+     * been resolved in their scope.
+     */
+    get<T>(identifier: Identifier<T>, parameters?: ParameterRecord): T;
+    /**
+     * Get all instances bound to the identifier.
+     * @param identifier The identifier to retrieve all bound instances for.
+     */
     getAll<T>(identifier: Identifier<T>): T[];
+    /**
+     * Checks to see if the identifier is known to the container.
+     * @param identifier The identifier to check.
+     */
     has(identifier: Identifier): boolean;
 }
 /**
@@ -41,6 +62,9 @@ export interface Context extends ServiceLocator {
      * To get scoped items, use Context.get() and Context.getAll()
      */
     readonly container: Container;
+    /**
+     * Injection parameters passed to this request.
+     */
     readonly parameters: ParameterRecord;
 }
 /**